+++ src/utils/variantOptions/index.ts
@@ -1,0 +1,4 @@
+export { filterAvailableOptions } from './filterAvailableOptions';
+export { findVariant } from './findVariant';
+export { getDownstreamAxisCodes } from './getDownstreamAxisCodes';
+export type { VariantAttribute, VariantLike } from './types';